Added Contact class with accessors and Array List of Contact
Contact class has 4 fields - name, phone, email and imageURL.  The
first are Strings and the last is a Drawable object.

I’ve added an email verifier and a String->Drawable converter at the
bottom of the class.  And overrided the toString.

Now working on the Create Contact layout to start passing the contact
object to and from, and to use ImageView rather than ImageButton for
displaying images.